Upgrade all dependencies
Also pin shift-codegen to an exact version, because the upgrade had
renamed an internal file that we hackishly require. Since we rely on
non-public parts of the package we have to pin it to avoid eslump
randomly breaking. Fixes #5 and supersedes #6.

Also pin all devDependencies that are part of the linting. I don’t want
them breaking for new contributors. The rest of the devDependencies are
for the examples; there it can be nice to allow later versions for
testing.